requirements.txt
flask
opencv-python
pandas
pillow
torch
numpy
Flask-SocketIO
eventlet